move EP create step button
